requirements.txt
adium-theme-ubuntu==0.3.4
args==0.1.0
attrs==15.2.0
autobahn==0.10.3
backports-abc==0.5
backports.shutil-get-terminal-size==1.0.0
backports.weakref==1.0rc1
beautifulsoup4==4.6.0
BitTorrent==3.4.2
bleach==1.5.0
blinker==1.3
boto==2.48.0
bs4==0.0.1
bz2file==0.98
certifi==2017.7.27.1
chardet==3.0.4
click==6.7
clint==0.5.1
configparser==3.5.0
cryptography==1.2.3
cycler==0.10.0
cymem==1.31.2
cytoolz==0.8.2
debtags==2.0
decorator==4.0.11
defer==1.0.6
dill==0.2.7.1
dirspec==13.10
docutils==0.12
duplicity==0.7.6
entrypoints==0.2.3
enum34==1.1.2
Flask==1.0.2
Flask-Uploads==0.2.1
floyd-cli==0.10.5
ftfy==4.4.3
funcsigs==1.0.2
functools32==3.2.3.post2
futures==3.0.5
gensim==2.3.0
genson==0.2.1
gunicorn==19.8.1
h5py==2.7.0
html5lib==0.9999999
httplib2==0.9.1
idna==2.6
ipaddress==1.0.16
ipykernel==4.6.1
ipython==5.4.1
ipython-genutils==0.2.0
ipywidgets==6.0.0
itsdangerous==0.24
Jinja2==2.10
jsonpickle==0.9.6
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.1.0
jupyter-console==5.1.0
jupyter-core==4.3.0
Keras==2.0.6
Kivy==1.10.0
lockfile==0.12.2
lxml==3.5.0
lz4==0.7.0
Markdown==2.6.8
MarkupSafe==1.0
marshmallow==2.13.6
matplotlib==2.0.2
mistune==0.7.4
mock==2.0.0
mpi4py==1.3.1
msgpack-python==0.4.6
murmurhash==0.26.4
nbconvert==5.2.1
nbformat==4.3.0
ndg-httpsclient==0.4.0
nose==1.3.7
notebook==5.0.0
numpy==1.13.1
oauthlib==1.0.3
oneconf==0.3.9
PAM==0.4.2
pandas==0.20.2
pandocfilters==1.4.1
pathlib==1.0.1
pathlib2==2.3.0
pbr==3.1.1
pexpect==4.2.1
pickleshare==0.7.4
Pillow==3.1.2
pipenv==11.10.4
piston-mini-client==0.7.5
plac==0.9.6
preshed==1.0.0
prompt-toolkit==1.0.14
protobuf==3.3.0
ptyprocess==0.5.2
pyasn1==0.1.9
pyasn1-modules==0.0.7
pycrypto==2.6.1
pycups==1.9.73
pygame==1.9.3
Pygments==2.2.0
pygobject==3.20.0
PyJWT==1.3.0
pyOpenSSL==0.15.1
pyparsing==2.2.0
pyserial==3.0.1
python-apt==1.1.0b1+ubuntu0.16.4.1
python-dateutil==2.6.0
python-debian==0.1.27
python-resize-image==1.1.11
python-snappy==0.5
pytz==2017.2
pyxdg==0.25
PyYAML==3.12
pyzmq==16.0.2
qtconsole==4.3.0
regex==2017.7.11
requests==2.18.4
requests-toolbelt==0.8.0
roman==2.0.0
scandir==1.5
scikit-learn==0.18.2
scipy==0.19.1
seaborn==0.7.1
service-identity==16.0.0
shortuuid==0.5.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.10.0
sklearn==0.0
smart-open==1.5.3
software-center-aptd-plugins==0.0.0
spacy==1.9.0
subprocess32==3.2.7
tabulate==0.7.7
tensorflow==1.2.1
termcolor==1.1.0
terminado==0.6
testpath==0.3.1
Theano==0.9.0
thinc==6.5.2
toolz==0.8.2
torch==0.1.12.post2
torchvision==0.1.8
tornado==4.5.1
tqdm==4.15.0
traitlets==4.3.2
trollius==2.0.1
Twisted==16.0.0
txaio==1.0.0
ubuntuone-client-data==14.4
ujson==1.35
unity-lens-photos==1.0
urllib3==1.22
virtualenv==15.2.0
virtualenv-clone==0.3.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==2.0.0
wrapt==1.10.10
youtube-dl==2016.2.22
zope.interface==4.1.3
